Add Master Scene and music
