--- Assets/CloudFloater.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 609e261031066ad4d8db5030bea7641a
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
