--- Assets/MillWingsRotator.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: e9cc4a5896740654dbabaa0d6b9b6a7e
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
