+++ Assets/Scripts/Animations.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 0b44c7b34263095458fc34ecf6593d30
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
