+++ Assets/Scripts/Managers/SceneIndexes.cs
@@ -1,0 +1,6 @@
+﻿public enum SceneIndexes
+{
+    MASTER = 0,
+    MAIN_MENU = 1,
+    GAME = 2,
+}
