+++ Assets/Scripts/Managers/SceneIndexes.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: d755ba33240b42242a8cbad53b8a70a2
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
