+++ Assets/Scripts/Managers/StartGameControl.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 281030ada5fbdda4197098b83909c4b8
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
